+++ app/javascript/selectors/accountSelectors.js
@@ -1,0 +1,5 @@
+export const selectCounty = state => {
+  return state.fetchAccount.account
+    ? state.fetchAccount.account.account.county_name
+    : '';
+};
