+++ app/javascript/selectors/userListSelector.js
@@ -3,12 +3,6 @@ export const selectUserRecords = state => {
   return usersObject ? usersObject.records : [];
 };
 
-export const selectCounty = state => {
-  return state.fetchAccount.account
-    ? state.fetchAccount.account.account.county_name
-    : '';
-};
-
 export const isLoading = state => {
   return state.fetchUserList.fetching || false;
 };
