+++ app/javascript/selectors/userListSelector.test.js
@@ -64,35 +64,4 @@ describe('selectors', () => {
       expect(isLoading(state)).toEqual(false);
     });
   });
-
-  describe('#selectCounty', () => {
-    it('selects the county when availble', () => {
-      const state = {
-        fetchAccount: {
-          account: {
-            XHRStatus: 'ready',
-            account: {
-              user: 'RACFID',
-              staff_id: '0X5',
-              county_code: '20',
-              county_name: 'Sacramento',
-            },
-          },
-        },
-      };
-      expect(selectCounty(state)).toEqual('Sacramento');
-    });
-
-    it('display empty string when county is not known', () => {
-      const state = {
-        fetchAccount: {
-          XHRStatus: 'ready',
-          user: 'RACFID',
-          staff_id: '0X5',
-          county_code: '20',
-        },
-      };
-      expect(selectCounty(state)).toEqual('');
-    });
-  });
 });
